[#285] Move load function to util
