+++ frontend/src/pages/manage-roadmap/CurrentRoadmap.jsx
@@ -12,6 +12,7 @@ import {
 import { api, store } from "../../lib";
 import { RiPencilFill, RiDeleteBinFill, RiPrinterFill } from "react-icons/ri";
 import { useNotification } from "../../util";
+import { handleLoad } from "../../util/common";
 import moment from "moment";
 
 const CurrentRoadmap = ({ setCurrentTab, setEditDatapoint }) => {
@@ -131,29 +132,6 @@ const CurrentRoadmap = ({ setCurrentTab, setEditDatapoint }) => {
       document.title = "ISCO";
     }, 1000);
   }
-
-  const handleLoad = (event) => {
-    const iframe = event.target;
-    if (iframe?.contentDocument) {
-      let css = "@page {";
-      css += "size: 210mm 297mm; margin: 15mm;";
-      css += "}";
-      css +=
-        "* { -webkit-print-color-adjust: exact !important; color-adjust: exact !important; }";
-      const style = document.createElement("style");
-      style.type = "text/css";
-      style.media = "print";
-      if (style.styleSheet) {
-        style.styleSheet.cssText = css;
-      } else {
-        style.appendChild(document.createTextNode(css));
-      }
-      const head = iframe.contentDocument.head;
-      if (head) {
-        head.appendChild(style);
-      }
-    }
-  };
 
   const columns = [
     {
